retui: code cleanup
organize code in separate files
__init__.py was cluttered
imports were a mess
