+++ print_tests/functional/bash_wrap.py
@@ -5,7 +5,7 @@ import subprocess  # nosec B404
 import time
 from pathlib import Path
 
-from retui import App, helper
+from retui import App, json_loader
 
 PROCESS_WRAP = None
 
@@ -157,7 +157,7 @@ def test(handle_sigint=True, demo_time_s=None, title=None):
             data = json.load(f)
             for widget in data["widgets"]:
                 print(widget)
-        app = helper.app_from_json(filename, globals(), app_dict={"post": sample_app_wrap})
+        app = json_loader.app_from_json(filename, globals(), app_dict={"post": sample_app_wrap})
         app.handle_sigint = handle_sigint
         app.demo_mode(demo_time_s)
         app.run()
