+++ print_tests/functional/json_app_wrap.py
@@ -3,7 +3,7 @@ import json
 import subprocess  # nosec B404
 from pathlib import Path
 
-from retui import App, helper
+from retui import App, json_loader
 
 PROCESS_WRAP = None
 
@@ -95,7 +95,7 @@ def test(handle_sigint=True, demo_time_s=None, title=None):
             data = json.load(f)
             for widget in data["widgets"]:
                 print(widget)
-        app = helper.app_from_json(filename, globals(), app_dict={"post": sample_app_wrap})
+        app = json_loader.app_from_json(filename, globals(), app_dict={"post": sample_app_wrap})
         app.handle_sigint = handle_sigint
         app.demo_mode(demo_time_s)
         app.run()
