+++ print_tests/functional/sample_main.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 import retui
 from retui.default_themes import DefaultThemes, DefaultThemesType
+from retui.enums import Dock, TextAlign
 from retui.widgets import Pane, TextBox
 
 
@@ -9,23 +10,21 @@ def test(handle_sigint=True, demo_time_s=None, title=None):
     app.title = title
     app.color_mode()
 
-    widget = TextBox(app=app, x=0, y=0, height=4, width=20, dock=retui.Dock.NONE)
+    widget = TextBox(app=app, x=0, y=0, height=4, width=20, dock=Dock.NONE)
     widget.text = "Test"
     app.add_widget(widget)
 
-    widget = Pane(app=app, x=2, y=8, height=4, width=8, dock=retui.Dock.NONE)
+    widget = Pane(app=app, x=2, y=8, height=4, width=8, dock=Dock.NONE)
     widget.title = "Little Pane"
     app.add_widget(widget)
 
-    pane = Pane(app=app, x=11, y=8, height=5, width=40, dock=retui.Dock.NONE)
+    pane = Pane(app=app, x=11, y=8, height=5, width=40, dock=Dock.NONE)
     pane.title = "Bigger Pane"
     app.add_widget(pane)
 
     test_color = retui.TerminalColor(retui.Color(13, retui.ColorBits.Bit8), retui.Color(7, retui.ColorBits.Bit8))
 
-    widget = TextBox(
-        app=app, x=0, y=0, height=3, width=10, dock=retui.Dock.NONE, text_align=retui.TextAlign.BottomRight
-    )
+    widget = TextBox(app=app, x=0, y=0, height=3, width=10, dock=Dock.NONE, text_align=TextAlign.BottomRight)
     widget.text = "Sample text in pane"
     widget.border_from_str(DefaultThemes.get_theme_border_str(DefaultThemesType.DOUBLE_TOP))
     widget.set_color(test_color)
@@ -37,8 +36,8 @@ def test(handle_sigint=True, demo_time_s=None, title=None):
         y=0,
         height=3,
         width=25,
-        dock=retui.Dock.NONE,
-        text_align=retui.TextAlign.BottomRight,
+        dock=Dock.NONE,
+        text_align=TextAlign.BottomRight,
     )
     widget.text = "TextBox without borders"
     widget.borderless = True
