+++ print_tests/functional/sample_main_fill_with_widgets.py
@@ -1,4 +1,5 @@
 import retui
+from retui.enums import DimensionsFlag, Dock, TextAlign
 from retui.widgets import Pane, TextBox
 
 
@@ -13,8 +14,8 @@ def test(handle_sigint=True, demo_time_s=None, title=None):
         y=0,
         height=80,
         width=80,
-        dock=retui.Dock.FILL,
-        dimensions=retui.DimensionsFlag.Fill,
+        dock=Dock.FILL,
+        dimensions=DimensionsFlag.Fill,
     )
     # dimensions should be ignored for Fill
     pane.title = "Test"
@@ -26,11 +27,11 @@ def test(handle_sigint=True, demo_time_s=None, title=None):
             y=0,
             height=10,
             width=20,
-            dock=retui.Dock.BOTTOM,
-            dimensions=retui.DimensionsFlag.FillWidth,
+            dock=Dock.BOTTOM,
+            dimensions=DimensionsFlag.FillWidth,
             text="The pane has 10 height.\nAnd is docked to BOTTOM so it should"
             " create nice bar at bottom.\nText alignment is Middle Right\n",
-            text_align=retui.TextAlign.MiddleRight,
+            text_align=TextAlign.MiddleRight,
         )
     )
 
@@ -41,11 +42,11 @@ def test(handle_sigint=True, demo_time_s=None, title=None):
             y=0,
             height=4,
             width=20,
-            dock=retui.Dock.RIGHT,
-            dimensions=retui.DimensionsFlag.Absolute,
+            dock=Dock.RIGHT,
+            dimensions=DimensionsFlag.Absolute,
             text="The pane has 20 width.\nIt is docked to the RIGHT so"
             "so this should go above bottom bar.\nText alignment is Middle Right\n",
-            text_align=retui.TextAlign.MiddleRight,
+            text_align=TextAlign.MiddleRight,
         )
     )
