+++ src/retui/input_handling/windows.py
@@ -1,0 +1,47 @@
+import ctypes
+import ctypes.wintypes
+
+
+class COORD(ctypes.Structure):
+    _fields_ = [("X", ctypes.wintypes.SHORT), ("Y", ctypes.wintypes.SHORT)]
+
+
+class KEY_EVENT_RECORD_Char(ctypes.Union):
+    _fields_ = [
+        ("UnicodeChar", ctypes.wintypes.WCHAR),
+        ("AsciiChar", ctypes.wintypes.CHAR),
+    ]
+
+
+class KEY_EVENT_RECORD(ctypes.Structure):
+    _fields_ = [
+        ("bKeyDown", ctypes.wintypes.BOOL),
+        ("wRepeatCount", ctypes.wintypes.WORD),
+        ("wVirtualKeyCode", ctypes.wintypes.WORD),
+        ("wVirtualScanCode", ctypes.wintypes.WORD),
+        ("uChar", KEY_EVENT_RECORD_Char),
+        ("dwControlKeyState", ctypes.wintypes.DWORD),
+    ]
+
+
+class MOUSE_EVENT_RECORD(ctypes.Structure):
+    _fields_ = [
+        ("dwMousePosition", COORD),
+        ("dwButtonState", ctypes.wintypes.DWORD),
+        ("dwControlKeyState", ctypes.wintypes.DWORD),
+        ("dwEventFlags", ctypes.wintypes.DWORD),
+    ]
+
+
+class INPUT_RECORD_Event(ctypes.Union):
+    _fields_ = [
+        ("KeyEvent", KEY_EVENT_RECORD),
+        ("MouseEvent", MOUSE_EVENT_RECORD),
+        ("WindowBufferSizeEvent", COORD),
+        ("MenuEvent", ctypes.c_uint),
+        ("FocusEvent", ctypes.c_uint),
+    ]
+
+
+class INPUT_RECORD(ctypes.Structure):
+    _fields_ = [("EventType", ctypes.wintypes.WORD), ("Event", INPUT_RECORD_Event)]
